Rename `handles.py` to `tdb_handles.py`.
